chore: move parser to config
